Trivial code cleanup
Just for the heck of it, I edited the Makefile and compiled with:

  gcc -W -Wall -pedantic

to see if there were any warnings or errors. Sure enough it exposed
several warnings, so I fixed some of those along the way. (I am in
the process of getting to know Atari 800 and I suppose your atr
utility can be useful on Linux.)